refactor: eliminate API import indirection and backward compatibility shim
- Remove deprecated api/api.js backward compatibility shim
- Update all 14 files to import directly from modular API structure
- eventSyncService.js → imports from api/events.js
- useSignInOut.js → imports from api/terms.js and api/flexiRecords.js
- auth.js → imports from api/auth.js
- All other files updated to use direct, explicit imports

Benefits:
- Eliminates unnecessary indirection in self-contained mobile app
- Makes import dependencies explicit and clear
- Simplifies codebase by removing roundabout file routing
- No backward compatibility needed since all users get new version
