+++ src/features/auth/services/auth.js
@@ -501,7 +501,7 @@ export async function fetchUserInfoFromAPI() {
     }
     
     try {
-      const apiModule = await import('../../../shared/services/api/api.js');
+      const apiModule = await import('../../../shared/services/api/api/auth.js');
       
       if (typeof apiModule.getStartupData !== 'function') {
         throw new Error('getStartupData function not available');
