+++ src/features/movements/hooks/useSectionMovements.js
@@ -5,7 +5,7 @@ import {
   getVikingSectionMoversData,
 } from '../../events/services/flexiRecordService.js';
 import { getToken } from '../../../shared/services/auth/tokenService.js';
-import { fetchMostRecentTermId } from '../../../shared/services/api/api.js';
+import { fetchMostRecentTermId } from '../../../shared/services/api/api/terms.js';
 import logger, { LOG_CATEGORIES } from '../../../shared/services/utils/logger.js';
 
 export default function useSectionMovements() {
