+++ src/features/sections/components/SectionsList.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { getListOfMembers } from '../../../shared/services/api/api.js';
+import { getListOfMembers } from '../../../shared/services/api/api/members.js';
 import { getToken } from '../../../shared/services/auth/tokenService.js';
 import { MemberDetailModal, MedicalDataPill } from '../../../shared/components/ui';
 import LoadingScreen from '../../../shared/components/LoadingScreen.jsx';
