+++ src/features/young-leaders/components/YoungLeadersPage.jsx
@@ -4,7 +4,8 @@ import LoadingScreen from '../../../shared/components/LoadingScreen.jsx';
 import databaseService from '../../../shared/services/storage/database.js';
 import logger, { LOG_CATEGORIES } from '../../../shared/services/utils/logger.js';
 import MainNavigation from '../../../shared/components/layout/MainNavigation.jsx';
-import { getListOfMembers, getTerms } from '../../../shared/services/api/api.js';
+import { getListOfMembers } from '../../../shared/services/api/api/members.js';
+import { getTerms } from '../../../shared/services/api/api/terms.js';
 import { getToken } from '../../../shared/services/auth/tokenService.js';
 import { UnifiedStorageService } from '../../../shared/services/storage/unifiedStorageService.js';
 import { isDemoMode } from '../../../config/demoMode.js';
