+++ src/shared/services/api/index.js
@@ -1,3 +1,2 @@
-// API services barrel export  
-export * from './api.js';
-// API directory with specific endpoints will be imported from ./api/ subdirectory
+// API services barrel export from modular structure
+export * from './api/index.js';
